Initial captcha support commit
