+++ lib/perimeterx.js
@@ -1,9 +1,6 @@
 'use strict';
 
-const pxCookie = require('./utils/pxcookie');
 const pxClient = require('./utils/pxclient');
-const pxApi = require('./utils/pxapi');
-const pxutil = require('./utils/pxutil');
 const pxlogger = require('./utils/pxlogger');
 
 /**
@@ -11,7 +8,6 @@ const pxlogger = require('./utils/pxlogger');
  * Version 1.0 Published 12 May 2016
  */
 module.exports.init = initPXModule;
-module.exports.verify = verifyUserScore;
 module.exports.middleware = pxMiddleware;
 
 /**
@@ -26,62 +22,6 @@ function initPXModule(params) {
 }
 
 /**
- * verifyUserScore - Verify function, evaluate score by cookie and s2s and make the return an action.
- *
- * @param {Object} pxCtx - current request context.
- *        {string} pxCtx.cookie - user's px cookie.
- *        {string} pxCtx.ip - user's ip address.
- *        {Array} pxCtx.headers - array of user's request headers in a name value format. (example: [{name: 'User-Agent', value: 'PhantomJS'}]
- *        {string} pxCtx.uri - current request uri
- * @param {Function} callback - callback function.
- */
-function verifyUserScore(pxCtx, callback) {
-    const pxconfig = require('./pxconfig').conf();
-    const cookie = pxCtx._px;
-    const ip = pxCtx.ip;
-    const headers = pxCtx.headers;
-    const uri = pxCtx.uri;
-
-    try {
-        if (!ip || !headers || !uri) {
-            pxlogger.error('perimeterx score evaluation failed. bad parameters.');
-            return callback(pxconfig.SCORE_EVALUATE_ACTION.PASS_TRAFFIC);
-        }
-        const userAgent = headers['user-agent'] || headers['User-Agent'] || 'none';
-
-        let action = pxCookie.evalCookie(cookie, ip, userAgent, pxCtx);
-        /* score did not cross threshold - pass traffic */
-        if (action === pxconfig.SCORE_EVALUATE_ACTION.GOOD_SCORE) {
-            return callback(pxconfig.SCORE_EVALUATE_ACTION.PASS_TRAFFIC);
-        }
-
-        /* score crossed threshold - block traffic */
-        if (action === pxconfig.SCORE_EVALUATE_ACTION.BAD_SCORE) {
-            return callback(pxconfig.SCORE_EVALUATE_ACTION.COOKIE_BLOCK_TRAFFIC);
-        }
-
-        /* when no fallback to s2s call if cookie does not exist or failed on evaluation */
-        pxApi.evalByServerCall(pxCtx, action, (action) => {
-            if (action === pxconfig.SCORE_EVALUATE_ACTION.UNEXPECTED_RESULT) {
-                pxlogger.error('perimeterx score evaluation failed. unexpected error. passing traffic');
-                return callback(pxconfig.SCORE_EVALUATE_ACTION.PASS_TRAFFIC);
-            }
-
-            if (action === pxconfig.SCORE_EVALUATE_ACTION.GOOD_SCORE) {
-                return callback(pxconfig.SCORE_EVALUATE_ACTION.PASS_TRAFFIC);
-            }
-
-            if (action === pxconfig.SCORE_EVALUATE_ACTION.BAD_SCORE) {
-                return callback(pxconfig.SCORE_EVALUATE_ACTION.S2S_BLOCK_TRAFFIC);
-            }
-        });
-    } catch (e) {
-        pxlogger.error('perimeterx score evaluation failed. unexpected error. ' + e.message);
-        return callback(pxconfig.SCORE_EVALUATE_ACTION.PASS_TRAFFIC);
-    }
-}
-
-/**
  * pxMiddleware - middleware wrapper to score verification.
  *
  * @param {Object} req - HTTP Request.
@@ -90,11 +30,11 @@ function verifyUserScore(pxCtx, callback) {
  */
 function pxMiddleware(req, res, next) {
     const pxconfig = require('./pxconfig').conf();
+    const pxutil = require('./utils/pxutil');
 
     if (req.method !== 'GET' || pxutil.checkForStatic(req, pxconfig.STATIC_FILES_EXT)) {
         return next();
     }
-
     try {
         const pxCtx = {};
         pxCtx._px = req.cookies ? req.cookies['_px'] : '';
@@ -102,16 +42,23 @@ function pxMiddleware(req, res, next) {
         pxCtx.headers = req.headers;
         pxCtx.uri = req.originalUrl || '/';
         pxCtx.full_url = req.protocol + '://' + req.get('host') + req.originalUrl;
-        verifyUserScore(pxCtx, (action) => {
-            pxlogger.debug('score action ' + action);
-            if (action === pxconfig.SCORE_EVALUATE_ACTION.PASS_TRAFFIC) {
-                pxlogger.debug('sending page requested activity');
-                pxClient.sendToPerimeterX('page_requested', {}, pxCtx);
-                return next();
-            }
 
-            return pxBlock(pxCtx, action, pxconfig.BLOCK_HANDLER, req, res, next);
-        });
+        if (req.cookies && req.cookies['_pxCaptcha']) {
+            pxutil.verifyCaptcha(req.cookies['_pxCaptcha'], (err, result) => {
+                return pxBlock(pxCtx, action, pxconfig.BLOCK_HANDLER, req, res, next);
+            });
+        } else {
+            pxutil.verifyUserScore(pxCtx, (action) => {
+                pxlogger.debug('score action ' + action);
+                if (action === pxconfig.SCORE_EVALUATE_ACTION.PASS_TRAFFIC) {
+                    pxlogger.debug('sending page requested activity');
+                    pxClient.sendToPerimeterX('page_requested', {}, pxCtx);
+                    return next();
+                }
+
+                return pxBlock(pxCtx, action, pxconfig.BLOCK_HANDLER, req, res, next);
+            });
+        }
     } catch (e) {
         return next();
     }
@@ -162,6 +109,6 @@ function defaultBlockHtmlGenerator(pxCtx) {
     const ref_str = pxCtx.block_uuid || '';
     const full_url = pxCtx.full_url;
 
-    var html = '<html lang="en">\n<head>\n    <link type="text/css" rel="stylesheet" media="screen, print"\n          href="//fonts.googleapis.com/css?family=Open+Sans:300italic,400italic,600italic,700italic,800italic,400,300,600,700,800">\n    <meta charset="UTF-8">\n    <title>Access to This Page Has Been Blocked</title>\n    <style> p {\n        width: 60%;\n        margin: 0 auto;\n        font-size: 35px;\n    }\n\n    body {\n        background-color: #a2a2a2;\n        font-family: "Open Sans";\n        margin: 5%;\n    }\n\n    img {\n        widht: 180px;\n    }\n\n    a {\n        color: #2020B1;\n        text-decoration: blink;\n    }\n\n    a:hover {\n        color: #2b60c6;\n    } </style>\n    <style type="text/css"></style>\n</head>\n<body cz-shortcut-listen="true">\n<div><img\n        src="http://storage.googleapis.com/instapage-thumbnails/035ca0ab/e94de863/1460594818-1523851-467x110-perimeterx.png">\n</div>\n<span style="color: white; font-size: 34px;">Access to This Page Has Been Blocked</span>\n<div style="font-size: 24px;color: #000042;"><br> Access to ' + full_url + 'is blocked according to the site security policy.\n    <br> Your browsing behaviour fingerprinting made us think you may be a bot. <br> <br> This may happen as a result of\n    the following:\n    <ul>\n        <li>JavaScript is disabled or not running properly.</li>\n        <li>Your browsing behaviour fingerprinting are not likely to be a regular user.</li>\n    </ul>\n    To read more about the bot defender solution: <a href="https://www.perimeterx.com/bot-defender">https://www.perimeterx.com/bot-defender</a>\n    <br> If you think the blocking was done by mistake, contact the site administrator. <br> <br> </br>' + ref_str + '</div>\n</body>\n</html>'
+    var html = '<html lang="en">\n<head>\n    <link type="text/css" rel="stylesheet" media="screen, print"\n          href="//fonts.googleapis.com/css?family=Open+Sans:300italic,400italic,600italic,700italic,800italic,400,300,600,700,800">\n    <meta charset="UTF-8">\n    <title>Access to This Page Has Been Blocked</title>\n    <style> p {\n        width: 60%;\n        margin: 0 auto;\n        font-size: 35px;\n    }\n\n    body {\n        background-color: #a2a2a2;\n        font-family: "Open Sans";\n        margin: 5%;\n    }\n\n    img {\n        widht: 180px;\n    }\n\n    a {\n        color: #2020B1;\n        text-decoration: blink;\n    }\n\n    a:hover {\n        color: #2b60c6;\n    } </style>\n    <style type="text/css"></style>\n    <script src=\'https://www.google.com/recaptcha/api.js\'></script>\n    <script>\n        function handleCaptcha(response) {\n            var name = \'_pxCaptcha\';\n            var expiryUtc = new Date( Date.now() + 1000 * 60 ).toUTCString();\n            var cookieParts = [name, \'=\', response, \'; expires=\', expiryUtc, \'; path=/\'];\n            document.cookie = cookieParts.join(\'\');\n            location.reload();\n        }\n    </script>\n</head>\n<body cz-shortcut-listen="true">\n<div><img\n        src="http://storage.googleapis.com/instapage-thumbnails/035ca0ab/e94de863/1460594818-1523851-467x110-perimeterx.png">\n</div>\n<span style="color: white; font-size: 34px;">Access to This Page Has Been Blocked</span>\n<div style="font-size: 24px;color: #000042;"><br> Access to ' + full_url + 'is blocked according to the site security policy.\n    <br> Your browsing behaviour fingerprinting made us think you may be a bot. <br> <br> This may happen as a result of\n    the following:\n    <ul>\n        <li>JavaScript is disabled or not running properly.</li>\n        <li>Your browsing behaviour fingerprinting are not likely to be a regular user.</li>\n    </ul>\n    To read more about the bot defender solution: <a href="https://www.perimeterx.com/bot-defender">https://www.perimeterx.com/bot-defender</a>\n    <br> If you think the blocking was done by mistake, contact the site administrator. <br> <br><div class="g-recaptcha" data-sitekey="6Lcj-R8TAAAAABs3FrRPuQhLMbp5QrHsHufzLf7b" data-callback="handleCaptcha" data-theme="dark"></div>\n\n    <span style="font-size: 20px;">Block Reference: <span\n            style="color: #525151;">#' + ref_str + '</span></span></div>\n</body>\n</html>'
     return html;
 }
