+++ lib/pxconfig.js
@@ -1,13 +1,12 @@
 'use strict';
-
 const _ = require('lodash');
-const pxutil = require('./utils/pxutil');
+
 let PX_DEFAULT = {};
 let PX_INTERNAL = {};
 let PX_CONF = {};
 
-module.exports.init = init;
-module.exports.conf = conf;
+exports.init = init;
+exports.conf = conf;
 
 /* internal configurations */
 PX_INTERNAL.SERVER_HOST = 'https://collector.a.pxi.pub';
@@ -41,7 +40,7 @@ PX_INTERNAL.SCORE_EVALUATE_ACTION = {
 
 /* to be defined by the initiating user */
 PX_DEFAULT.PX_APP_ID = 'PX_APP_ID';
-PX_DEFAULT.API_TIMEOUT_MS = 1000;
+PX_DEFAULT.API_TIMEOUT_MS = 3000;
 PX_DEFAULT.BLOCKING_SCORE = 70;
 PX_DEFAULT.COOKIE_SECRET_KEY = 'cookie_secret_key';
 PX_DEFAULT.AUTH_TOKEN = 'auth_token';
@@ -54,18 +53,40 @@ PX_DEFAULT.BLOCK_HANDLER = '';
 PX_DEFAULT.MAX_BUFFER_LEN = 30;
 
 function init(params) {
-    PX_DEFAULT.PX_APP_ID = pxutil.configurationsOverriding(PX_DEFAULT, params, 'PX_APP_ID', 'pxAppId');
-    PX_DEFAULT.COOKIE_SECRET_KEY = pxutil.configurationsOverriding(PX_DEFAULT, params, 'COOKIE_SECRET_KEY', 'cookieSecretKey');
-    PX_DEFAULT.AUTH_TOKEN = pxutil.configurationsOverriding(PX_DEFAULT, params, 'AUTH_TOKEN', 'authToken');
-    PX_DEFAULT.PROXY_URL = pxutil.configurationsOverriding(PX_DEFAULT, params, 'PROXY_URL', 'proxy');
-    PX_DEFAULT.API_TIMEOUT_MS = pxutil.configurationsOverriding(PX_DEFAULT, params, 'API_TIMEOUT_MS', 'apiTimeoutMS');
-    PX_DEFAULT.BLOCK_HANDLER = pxutil.configurationsOverriding(PX_DEFAULT, params, 'BLOCK_HANDLER', 'blockHandler');
-    PX_DEFAULT.BLOCKING_SCORE = pxutil.configurationsOverriding(PX_DEFAULT, params, 'BLOCKING_SCORE', 'blockingScore');
-    PX_DEFAULT.IP_HEADER = pxutil.configurationsOverriding(PX_DEFAULT, params, 'IP_HEADER', 'ipHeader');
-    PX_DEFAULT.SEND_PAGE_ACTIVITIES = pxutil.configurationsOverriding(PX_DEFAULT, params, 'SEND_PAGE_ACTIVITIES', 'sendPageActivities');
-    PX_DEFAULT.DEBUG_MODE = pxutil.configurationsOverriding(PX_DEFAULT, params, 'DEBUG_MODE', 'debugMode');
-    PX_DEFAULT.MAX_BUFFER_LEN = pxutil.configurationsOverriding(PX_DEFAULT, params, 'MAX_BUFFER_LEN', 'maxBufferLength');
+    PX_DEFAULT.PX_APP_ID = configurationsOverriding(PX_DEFAULT, params, 'PX_APP_ID', 'pxAppId');
+    PX_DEFAULT.COOKIE_SECRET_KEY = configurationsOverriding(PX_DEFAULT, params, 'COOKIE_SECRET_KEY', 'cookieSecretKey');
+    PX_DEFAULT.AUTH_TOKEN = configurationsOverriding(PX_DEFAULT, params, 'AUTH_TOKEN', 'authToken');
+    PX_DEFAULT.PROXY_URL = configurationsOverriding(PX_DEFAULT, params, 'PROXY_URL', 'proxy');
+    PX_DEFAULT.API_TIMEOUT_MS = configurationsOverriding(PX_DEFAULT, params, 'API_TIMEOUT_MS', 'apiTimeoutMS');
+    PX_DEFAULT.BLOCK_HANDLER = configurationsOverriding(PX_DEFAULT, params, 'BLOCK_HANDLER', 'blockHandler');
+    PX_DEFAULT.BLOCKING_SCORE = configurationsOverriding(PX_DEFAULT, params, 'BLOCKING_SCORE', 'blockingScore');
+    PX_DEFAULT.IP_HEADER = configurationsOverriding(PX_DEFAULT, params, 'IP_HEADER', 'ipHeader');
+    PX_DEFAULT.SEND_PAGE_ACTIVITIES = configurationsOverriding(PX_DEFAULT, params, 'SEND_PAGE_ACTIVITIES', 'sendPageActivities');
+    PX_DEFAULT.DEBUG_MODE = configurationsOverriding(PX_DEFAULT, params, 'DEBUG_MODE', 'debugMode');
+    PX_DEFAULT.MAX_BUFFER_LEN = configurationsOverriding(PX_DEFAULT, params, 'MAX_BUFFER_LEN', 'maxBufferLength');
     PX_CONF = _.merge(PX_DEFAULT, PX_INTERNAL);
+}
+
+function configurationsOverriding(conf, params, defaultName, userInput) {
+    /* user did not override configuration */
+    if (!params[userInput]) {
+        return conf[defaultName];
+    }
+
+    /* for type mismatch, we'll use the default */
+    if (typeof conf[defaultName] != typeof params[userInput]) {
+        return conf[defaultName];
+    }
+
+    /* handling block handler overriding */
+    if (userInput == 'blockHandler' && params[userInput]) {
+        if (typeof params[userInput] == 'function') {
+            return params[userInput];
+        }
+        return '';
+    }
+
+    return params[userInput];
 }
 
 function conf() {
