+++ lib/utils/pxutil.js
@@ -1,12 +1,87 @@
 'use strict';
 
+const pxCookie = require('./pxcookie');
+const pxApi = require('./pxapi');
+const pxlogger = require('./pxlogger');
+
 /**
  * PerimeterX (http://www.perimeterx.com) NodeJS-Express SDK
  * Version 1.0 Published 12 May 2016
  */
+exports.verifyUserScore = verifyUserScore;
+exports.verifyCaptcha = verifyCaptcha;
 exports.formatHeaders = formatHeaders;
 exports.checkForStatic = checkForStatic;
 exports.configurationsOverriding = configurationsOverriding;
+
+/**
+ * verifyUserScore - Verify function, evaluate score by cookie and s2s and make the return an action.
+ *
+ * @param {Object} pxCtx - current request context.
+ *        {string} pxCtx.cookie - user's px cookie.
+ *        {string} pxCtx.ip - user's ip address.
+ *        {Array} pxCtx.headers - array of user's request headers in a name value format. (example: [{name: 'User-Agent', value: 'PhantomJS'}]
+ *        {string} pxCtx.uri - current request uri
+ * @param {Function} callback - callback function.
+ */
+function verifyUserScore(pxCtx, callback) {
+    const pxconfig = require('../pxconfig').conf();
+    const cookie = pxCtx._px;
+    const ip = pxCtx.ip;
+    const headers = pxCtx.headers;
+    const uri = pxCtx.uri;
+
+    try {
+        if (!ip || !headers || !uri) {
+            pxlogger.error('perimeterx score evaluation failed. bad parameters.');
+            return callback(pxconfig.SCORE_EVALUATE_ACTION.PASS_TRAFFIC);
+        }
+        const userAgent = headers['user-agent'] || headers['User-Agent'] || 'none';
+
+        let action = pxCookie.evalCookie(cookie, ip, userAgent, pxCtx);
+        /* score did not cross threshold - pass traffic */
+        if (action === pxconfig.SCORE_EVALUATE_ACTION.GOOD_SCORE) {
+            return callback(pxconfig.SCORE_EVALUATE_ACTION.PASS_TRAFFIC);
+        }
+
+        /* score crossed threshold - block traffic */
+        if (action === pxconfig.SCORE_EVALUATE_ACTION.BAD_SCORE) {
+            return callback(pxconfig.SCORE_EVALUATE_ACTION.COOKIE_BLOCK_TRAFFIC);
+        }
+
+        /* when no fallback to s2s call if cookie does not exist or failed on evaluation */
+        pxApi.evalByServerCall(pxCtx, action, (action) => {
+            if (action === pxconfig.SCORE_EVALUATE_ACTION.UNEXPECTED_RESULT) {
+                pxlogger.error('perimeterx score evaluation failed. unexpected error. passing traffic');
+                return callback(pxconfig.SCORE_EVALUATE_ACTION.PASS_TRAFFIC);
+            }
+
+            if (action === pxconfig.SCORE_EVALUATE_ACTION.GOOD_SCORE) {
+                return callback(pxconfig.SCORE_EVALUATE_ACTION.PASS_TRAFFIC);
+            }
+
+            if (action === pxconfig.SCORE_EVALUATE_ACTION.BAD_SCORE) {
+                return callback(pxconfig.SCORE_EVALUATE_ACTION.S2S_BLOCK_TRAFFIC);
+            }
+        });
+    } catch (e) {
+        pxlogger.error('perimeterx score evaluation failed. unexpected error. ' + e.message);
+        return callback(pxconfig.SCORE_EVALUATE_ACTION.PASS_TRAFFIC);
+    }
+}
+
+
+/**
+ * verifyUserScore - Verify function, evaluate ca.
+ *
+ * @param {string} pxCaptcha - captcha value.
+ * @param {Function} callback - callback function.
+ */
+function verifyCaptcha(pxCaptcha, callback) {
+    console.log(pxCaptcha);
+    return callback();
+}
+
 
 /**
  * formatHeaders - Build request headers in the server2server format.
@@ -32,28 +107,6 @@ function formatHeaders(headers) {
     }
 }
 
-
-function configurationsOverriding(conf, params, defaultName, userInput) {
-    /* user did not override configuration */
-    if (!params[userInput]) {
-        return conf[defaultName];
-    }
-
-    /* for type mismatch, we'll use the default */
-    if (typeof conf[defaultName] != typeof params[userInput]) {
-        return conf[defaultName];
-    }
-
-    /* handling block handler overriding */
-    if (userInput == 'blockHandler' && params[userInput]) {
-        if (typeof params[userInput] == 'function') {
-            return params[userInput];
-        }
-        return '';
-    }
-
-    return params[userInput];
-}
 
 /**
  * checkForStatic - check if the request destination is a static file.
